add REFERENCE_THRUST_TO_VERTICAL setting

+++ modules/interface/commander.h
@@ -44,7 +44,7 @@ void commanderWatchdog(void);
 uint32_t commanderGetInactivityTime(void);
 void commanderGetRPY(float* eulerRollDesired, float* eulerPitchDesired, float* eulerYawDesired);
 void commanderGetRPYType(RPYType* rollType, RPYType* pitchType, RPYType* yawType);
-void commanderGetThrust(uint16_t* thrust);
+void commanderGetThrust(float* thrust);
 void commanderGetAltHold(bool* altHold, bool* setAltHold, float* altHoldChange);
 
 #endif /* COMMANDER_H_ */
